# doc/api.yaml
openapi: 3.0.0
info:
  title: 'WASAText API'
  description: 'API for managing a messaging platform'
  version: 1.0.0
servers:                                              # Base URL where the API is
  - url: 'http://localhost:3000'
    description: 'Local Development Environment (Docker/Go)'
  - url: 'https://api.wasatext.com'
    description: 'Production Server'
security:                                             # Global security
  - BearerAuth: []
components:                                           # Reusable components
  schemas:                                            # Structures
    Error:
      description: 'Standard error'
      type: object
      properties:
        code:
          description: 'Integer code error'
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          readOnly: true
        message:
          description: 'Error text message'
          type: string
          pattern: '^.{0,500}$'                       # 0-500 chars
          minLength: 0
          maxLength: 500
          readOnly: true
      required: [code, message]
      additionalProperties: false                     # No other properties allowed
      example:
        code: 404
        message: 'The requested resource could not be found'
    Emoji:
      description: |
        Exactly one grapheme cluster, capped at 16 Unicode code points.
        Validation is structural rather than an emoji-catalogue lookup, so any one symbol is accepted.
      type: string
      pattern: '^.{1,16}$'                            # 1 symbol, at most 16 code points
      minLength: 1
      maxLength: 16
      example: '👍'
    Id:
      description: 'Standard unique identifier, only the canonical lowercase form is valid'
      type: string
      format: uuid                                    # Universally Unique Identifier
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
      minLength: 36
      maxLength: 36
      example: 'c56a4180-65aa-42ec-a945-5fd21dec0538'
...
    UserId:
      $ref: '#/components/schemas/Id'
    ChatId:
      $ref: '#/components/schemas/Id'
    MessageId:
      $ref: '#/components/schemas/Id'
    CommentId:
      $ref: '#/components/schemas/Id'
    Photo:
      description: 'Binary photo file'
      type: string
      format: binary
      pattern: '^.*$'                                 # Any binary content
      minLength: 1
      maxLength: 31457280                             # 30 Megabyte (30 * 1024 * 1024)
      example: 'the raw bytes of the photo'
    PhotoURL:
      description: |
        Photo URL, relative to the server: the client already knows the host it is talking to.
        The last segment is the id of the photo: a UUID saved with no extension,
        the content type is detected when the photo is served
      type: string                                    # Relative reference, not a whole URL
      pattern: '^/photos/[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
      minLength: 44
      maxLength: 44
      readOnly: true
      example: '/photos/c56a4180-65aa-42ec-a945-5fd21dec0538'
...
    ChatBase:                                         # Shape: what a chat carries whichever way it is asked for
      description: |
        What a chat carries whichever way it is asked for, and the root every chat shape grows from:
        no members, no messages and no preview, each of them added by the shape that answers with it
        chatType and name are left open to both kinds here, and narrowed by the kind axis below
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ChatId'
        chatType:                                     # The discriminator of the two kinds,
          description: 'Chat type'                    # narrowed to a single value by the kind axis
          type: string
          pattern: '^.{5,7}$'                         # 5-7 chars as 'group' or 'private'
          minLength: 5
          maxLength: 7
          enum: ['private', 'group']
          readOnly: true
          example: 'group'
        name:                                         # A GroupName for a group, a Username for a private chat,
          description: |                              # so what stands here is the wider of the two
            The name of the chat: the one a group owns,
            or the username of the other member of a private chat
          type: string
          minLength: 1
          maxLength: 100
        photo:                                        # The same for both kinds, so the kind axis narrows it in neither:
          $ref: '#/components/schemas/PhotoURL'       # a group owns its photo, a private chat borrows the other member's
      required: [id, chatType, name, photo]
...
    ChatWithMembers:                                  # Shape: the base + who belongs to the chat
      description: |
        A chat and its members, without the messages
        A chat is answered with its members only to a member of it, so the list always holds at least the caller
      allOf:                                          # Combines ChatBase with the modifications below
        - $ref: '#/components/schemas/ChatBase'       # Inherit everything from ChatBase
        - type: object                                # Override type to object
          properties:
            members:
              description: 'Members of the chat'
              type: array
              items:
                $ref: '#/components/schemas/UserId'
              minItems: 1                             # The wider of the two kinds, narrowed by the kind axis below
              maxItems: 100
              uniqueItems: true                       # One row per membership: a member cannot be held twice
          required: [members]                         # + id, chatType, name, photo already required in ChatBase
...
    GroupKind:                                        # Kind: what a group does not share with a pair
      description: |
        The narrowing that makes a chat a group, and nothing else
        It names no property the shape axis does not already carry, so it composes with every shape above:
        members is narrowed and never required here, so a shape that carries none is left untouched by it
      type: object
      properties:
        chatType:
          description: 'Chat type'
          type: string
          pattern: '^.{5}$'                           # 5 chars as 'group'
          minLength: 5
          maxLength: 5
          enum: ['group']
          readOnly: true
          example: 'group'
        name:                                         # The name the group owns
          $ref: '#/components/schemas/GroupName'
        members:                                      # Only the count narrows here, the items are the ones
          description: 'Members of the group'         # The shape this is combined with already carries
          type: array
          items:
            $ref: '#/components/schemas/UserId'
          minItems: 1                                 # A group answered with its members holds at least the caller
          maxItems: 100
...
    GroupWithMembers:                                 # Wire shape: addToGroup
      description: 'A group chat with its members, without the messages'
      allOf:
        - $ref: '#/components/schemas/ChatWithMembers'
        - $ref: '#/components/schemas/GroupKind'
...
    MessageBase:
      description: 'Base structure for messages and snippets'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/MessageId'
        date:
          description: 'Message creation date in UTC, with millisecond precision'
          type: string
          format: date-time
          example: '2025-11-04T10:32:00.000Z'
        user:
          $ref: '#/components/schemas/UserId'
        state:
          description: |
            Message state, computed and never stored:
            read: no member of the chat is left behind the message,
            received: at least one of them still is.
            A member falls behind a message when the message arrives, and catches up by opening the chat
            or by writing in it, since writing means having seen what is above.
            The sender needs no exception: sending catches it up to its own message.
            A member joins caught up, so it answers for nothing sent before it and a message
            already read never turns back; a member leaving takes its answer with it,
            so a message the leaver alone was holding becomes read.
            The value is a fact about the members and not about the caller,
            so it is the same for everybody looking at the chat
          type: string
          pattern: '^.{4,8}$'                         # 4-8 chars as 'read' or 'received'
          minLength: 4
          maxLength: 8
          enum: [received, read]
          example: 'read'
          # EXTRA: sent: message sent by the user but not received yet by all chat members
      required: [id, date, user, state]
    MessageText:
      description: 'Standard message text content'
      type: string
      pattern: '^.{1,5000}$'                          # 1-5000 chars
      minLength: 1
      maxLength: 5000
      example: 'Hello world!'
    Message:
      description: 'Standard message'
      allOf:                                          # Combines MessageBase with the modifications below
        - $ref: '#/components/schemas/MessageBase'    # Inherit everything from MessageBase
        - type: object                                # Override type to object
          properties:
            content:
              description: 'Content of the message, text and/or photo'
              type: object
              properties:
                text:
                  $ref: '#/components/schemas/MessageText'
                photo:
                  $ref: '#/components/schemas/PhotoURL'
              minProperties: 1                        # At least one of text/photo is present
            comments:
              description: 'The reactions on this message, empty list if none'
              type: array
              items:
                $ref: '#/components/schemas/Comment'
              minItems: 0
              maxItems: 1000
              uniqueItems: true                       # One comment per user, and one id per comment
          required: [content, comments]               # + id, date, user, state already required in MessageBase
...
    Comment:
      description: 'Standard comment to a message'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/CommentId'
        user:
          $ref: '#/components/schemas/UserId'
        emoji:
          $ref: '#/components/schemas/Emoji'
      required: [id, user, emoji]
      additionalProperties: false                     # No other properties allowed
  securitySchemes:                                    # Authentication
    BearerAuth:
      description: 'Authentication token (JWT)'       # JSON Web Token
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ParamChatId:
      name: chatId
      description: 'Chat unique identifier'
      in: path
      schema:
        $ref: '#/components/schemas/ChatId'
      required: true
    ParamGroupId:                                     # Same id space as ParamChatId, one kind of chat less:
      name: groupId                                   # under /groups the id must name a group, and a private chat is a 404
      description: 'Group unique identifier'
      in: path
      schema:
        $ref: '#/components/schemas/ChatId'
      required: true
    ParamMessageId:
      name: messageId
      description: 'Message unique identifier'
      in: path
      schema:
        $ref: '#/components/schemas/MessageId'
      required: true

    # EXTRA: add cursor based pagination
  responses:
...
    NoContent:
      description: '204 No Content'
    
    # Client errors
    BadRequest:
      description: '400 Bad Request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400
            message: 'The data provided is not valid. Please check constraints and required fields'
    Unauthorized:
      description: '401 Unauthorized'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 401
            message: 'Authentication required. Please include a valid access token'
    Forbidden:
      description: '403 Forbidden'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 403
            message: 'You do not have permission to access this resource'
    NotFound:
      description: '404 Not Found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 404
            message: 'The requested resource was not found'

    #MethodNotAllowed:
    #  description: '405 Method Not Allowed'
    #  content:
    #    application/json:
    #      schema:
    #        $ref: '#/components/schemas/Error'
    #      example:
    #        code: 405
    #        message: 'The HTTP method used is not supported for this resource'

    # Server errors
    InternalServerError:
      description: '500 Internal Server Error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 500
            message: 'An internal server error occurred'
...
paths:                                                # Rest endpoints
...
  /chats/{chatId}/messages:
    post:
      operationId: 'sendMessage'
      description: |
        Writes a message in a chat the authenticated user is a member of.
        Both kinds (private chat and group) are written here, as they are read under /chats/{chatId}.
        A message carries text, a photo, or both, so the file part is optional,
        and a text of spaces alone counts as no text: a message that would carry neither is a bad request.
        Sending catches the sender up with the chat, since writing in it means having seen what is above.
        In a chat the sender is alone in there is nobody to be behind, so the message is born 'read'.
        A chat holds up to 10000 messages, and one that is full is a bad request and not a conflict.
      summary: 'Send a new message'
      tags: ['PrivateChat', 'Group']
      parameters:
        - $ref: '#/components/parameters/ParamChatId'
      requestBody:
        description: 'Message payload containing text and/or a binary photo file'
        content:
          multipart/form-data:
            schema:
              description: 'Object containing the message content'
              type: object
              properties:
                text:
                  $ref: '#/components/schemas/MessageText'
                photoFile:
                  $ref: '#/components/schemas/Photo'
              minProperties: 1                        # at least one of text/photo is present
        required: true
      responses:
        '201':
          description: 'Message sent successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                id: '3fa85f64-5717-4562-b3fc-2c963f66afa6'
                content:
                  text: 'Hello, how are you?'
                date: '2025-11-15T18:45:00.000Z'
                user: 'cf6a4180-65aa-42ec-a945-5fd21dec0538'
                state: 'received'                     # Nobody else has seen it yet
                comments: []                          # A message is born with no reaction
        '400':                                        # Empty content, broken text, bad photo,
          $ref: '#/components/responses/BadRequest'   # a body over the limit, or a chat already full
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':                                        # Not a member of the chat
          $ref: '#/components/responses/Forbidden'
        '404':                                        # No chat owns that id
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /chats/{chatId}/messages/forwards:
    post:
      operationId: 'forwardMessage'
      description: |
        Copies an existing message into a destination chat.
        The authenticated user must be a member of both the source chat and the destination chat.
        The source message is left untouched. The forwarded copy has a new id and date,
        identifies the authenticated user as its sender, carries the source text and/or photo, and has no comments.
        Forwarding catches the sender up with the destination chat, since writing in it means having seen what is above.
        Its state is therefore computed like that of a newly sent message:
        it is 'read' when no other destination member is behind it, and 'received' otherwise.
        A destination chat holds up to 10000 messages, and one that is full is a bad request.
      summary: 'Forward a message'
      tags: ['PrivateChat', 'Group']
      parameters:
        - $ref: '#/components/parameters/ParamChatId'
      requestBody:
        description: 'Payload containing the unique identifier of the message to be forwarded'
        content:
          application/json:
            schema:
              description: 'Object containing the ID of the message to forward'
              type: object
              additionalProperties: false
              properties:
                messageId:
                  $ref: '#/components/schemas/MessageId'
              required: [messageId]
        required: true
      responses:
        '201':
          description: 'Message forwarded successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                id: '4fa85f64-5717-4562-b3fc-2c963f66afa7'
                content:
                  text: 'Hello, how are you?'
                date: '2025-11-15T19:00:00.000Z'
                user: 'cf6a4180-65aa-42ec-a945-5fd21dec0538'
                state: 'received'
                comments: []
        '400':                                        # Invalid destination id, invalid body/source id,
          $ref: '#/components/responses/BadRequest'   # or a destination chat that is already full
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':                                        # Not a member of the source or destination chat
          $ref: '#/components/responses/Forbidden'
        '404':                                        # No source message or destination chat owns the given id
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /chats/{chatId}/messages/{messageId}/comments/me: # current user unique comment
    parameters:
      - $ref: '#/components/parameters/ParamChatId'
      - $ref: '#/components/parameters/ParamMessageId'

    put:
      operationId: 'commentMessage'
      description: |
        Sets the authenticated user's reaction on a message in a chat they belong to.
        The message must belong to the chat named in the URL.
        Each member has at most one reaction on a message; another call updates that row's emoji while preserving its comment id.
        Commenting catches the caller up with the chat since writing in it means having seen what is above.
        The returned Message contains its unchanged content plus the complete updated comments list and its state after that catch-up. 
      summary: 'Comment a message'
      tags: ['PrivateChat', 'Group']
      requestBody:
        description: 'Payload containing the new single emoji to set as comment'
        content:
          application/json:
            schema:
              description: 'Object containing the emoji of the comment'
              type: object                            # not a Comment: id and user are readOnly,
              additionalProperties: false             # No other properties allowed by the contract
              properties:                             # the server knows both and the client cannot send them
                emoji:
                  $ref: '#/components/schemas/Emoji'
              required: [emoji]
        required: true
      responses:
        '200':
          description: 'Existing comment updated successfully; returns the message with its updated comments'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                id: '3fa85f64-5717-4562-b3fc-2c963f66afa6'
                content:
                  text: 'Hello, how are you?'
                date: '2025-11-15T18:45:00.000Z'
                user: 'cf6a4180-65aa-42ec-a945-5fd21dec0538'
                state: 'received'
                comments:
                  - id: 'c52a4180-65aa-42ec-a945-5fd21dec0538'
                    emoji: '👍'
                    user: '1a6a4180-65aa-42ec-a945-5fd21dec0538'
        '201':
          description: 'New comment created successfully; returns the message with its updated comments'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                id: '3fa85f64-5717-4562-b3fc-2c963f66afa6'
                content:
                  text: 'Hello, how are you?'
                date: '2025-11-15T18:45:00.000Z'
                user: 'cf6a4180-65aa-42ec-a945-5fd21dec0538'
                state: 'received'
                comments:
                  - id: 'c52a4180-65aa-42ec-a945-5fd21dec0538'
                    emoji: '👍'
                    user: '1a6a4180-65aa-42ec-a945-5fd21dec0538'
        '400':                                        # Invalid URL id, body or emoji
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':                                        # The message belongs to this chat, but the caller does not
          $ref: '#/components/responses/Forbidden'
        '404':                                        # No message owns the id, or it belongs to another chat
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      operationId: 'uncommentMessage'
      description: |
        Removes the authenticated user's own reaction from a message in a chat they belong to.
        The message must belong to the chat named in the URL, and the caller must already have a comment on the message in the URL.
        Other members' comments and the message itself are left untouched.
        A successful removal catches the caller up with the chat without moving lastReadDate backwards.
      summary: 'Remove my comment from a message'
      tags: ['PrivateChat', 'Group']
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '400':                                        # Invalid chatId or messageId
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':                                        # The message belongs to this chat, but the caller does not
          $ref: '#/components/responses/Forbidden'
        '404':                                        # Message/chat mismatch, missing message, or no caller comment
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /groups/{groupId}/members:
    post:
      operationId: 'addToGroup'
      description: |
        Adds one or more users to a group the authenticated user is a member of.
        A user already inside is ignored, so the same request can be repeated without failing,
        and an empty list is a request that adds nobody: it still answers for the group.
        A private chat borrows its two members from its pair and takes no more,
        so its id names nothing under this collection and is a 404 here.
      summary: 'Add users in a group'
      tags: ['Group']
      parameters:
        - $ref: '#/components/parameters/ParamGroupId'
      requestBody:
        description: 'List of the UserId of the new members to add to the group'
        content:
          application/json:
            schema:
              description: 'Payload for new members'
              type: object
              properties:
                members:
                  description: 'New members userId'
                  type: array
                  items:
                    $ref: '#/components/schemas/UserId'
                  minItems: 0                             # An empty list is a request that adds nobody
                  maxItems: 100                           # Better: group max size - group current size
                  uniqueItems: true                       # Ensure that each member could be once in the members list
        required: true
      responses:
        '200':
          description: 'New members added to the group successfully'
          content:
            application/json:
              schema:                                 # No messages list: adding members never reads them,
                $ref: '#/components/schemas/GroupWithMembers' # and no preview is derived from them
              example:
                id: '9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d'
                chatType: 'group'
                name: 'Computer Science Study Group'
                photo: '/photos/d41f8a72-3b6c-4e19-8a52-0c7f5b21e934'
                members: ['90b2c3d4-e5f6-42ec-a945-5fd21dec0538', '70b2c3d4-e5f6-42ec-a945-5fd21dec0538', '80b2c3d4-e5f6-42ec-a945-5fd21dec0538']
        '400':                                        # Malformed body, or the additions do not fit the group
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':                                        # Not a member of the group
          $ref: '#/components/responses/Forbidden'
        '404':                                        # No group owns that id, or one of the members does not exist
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...